requirements.txt
certifi==2020.4.5.1
coloredlogs==14.0
easydict==1.9
humanfriendly==8.2
opencv-python==4.2.0.34
Pillow==7.1.1
protobuf==3.11.3
PyYAML==5.3.1
scipy==1.3.1
six==1.14.0
tensorboardX==2.0
torch==1.5.1
torchvision==0.6.1
matplotlib==3.2.2
numpy=1.19.0
